ai/pyproject.toml
[project]
name = "smart-invoice-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.7",
    "google-genai>=1.62.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
